### 2021-05-05 #3254 @whedon
[PRE REVIEW]: DataLad: distributed system for joint management of code, data, and their relationship
**Submitting author:** @yarikoptic (<a href="http://orcid.org/0000-0003-3456-2493">Yaroslav Halchenko</a>)
**Repository:** <a href="https://github.com/datalad/datalad" target ="_blank">https://github.com/datalad/datalad</a>
**Version:** 0.14.3
