### 2021-05-05 878c2ea
ENH: clone: Accept arbitrary git clone options
git-clone has several options that are useful to expose, including
--branch for specifying a branch or tag to clone as well as --filter
for partial clones and --sparse for sparse checkouts.  (However, some
options, like --single-branch, don't play well with annex
repositories.)

Follow the approach of datalad-create, and extend datalad-clone to
take arbitrary git-clone options.

Re: #2109
Closes #4035
Re: #5286
