+++ datalad/core/distributed/clone.py
@@ -9,6 +9,7 @@
 """Plumbing command for dataset installation"""
 
 
+from argparse import REMAINDER
 import logging
 import re
 from os.path import expanduser
@@ -196,6 +197,18 @@ class Clone(Interface):
             doc="""path to clone into.  If no `path` is provided a
             destination path will be derived from a source URL
             similar to :command:`git clone`"""),
+        git_clone_opts=Parameter(
+            args=("git_clone_opts",),
+            metavar='GIT CLONE OPTIONS',
+            nargs=REMAINDER,
+            doc="""[PY: A list of command line arguments PY][CMD: Options CMD]
+            to pass to :command:`git clone`. [CMD: Any argument specified after
+            SOURCE and the optional PATH will be passed to git-clone. CMD] Note
+            that not all options will lead to viable results. For example
+            '--single-branch' will not result in a functional annex repository
+            because both a regular branch and the git-annex branch are
+            required. Note that a version in a RIA URL takes precedence over
+            '--branch'."""),
         description=location_description,
         reckless=reckless_opt,
     )
@@ -208,7 +221,8 @@ class Clone(Interface):
             path=None,
             dataset=None,
             description=None,
-            reckless=None):
+            reckless=None,
+            git_clone_opts=None):
         # did we explicitly get a dataset to install into?
         # if we got a dataset, path will be resolved against it.
         # Otherwise path will be resolved first.
@@ -296,6 +310,7 @@ class Clone(Interface):
             description,
             result_props,
             cfg=None if ds is None else ds.config,
+            clone_opts=git_clone_opts
         )
 
         # TODO handle any 'version' property handling and verification using a
@@ -325,7 +340,8 @@ def clone_dataset(
         description=None,
         result_props=None,
         cfg=None,
-        checkout_gitsha=None):
+        checkout_gitsha=None,
+        clone_opts=None):
     """Internal helper to perform cloning without sanity checks (assumed done)
 
     This helper does not handle any saving of subdataset modification or adding
@@ -354,6 +370,10 @@ def clone_dataset(
       did not obtain the commit object. Should the checkout of the target commit
       cause a detached HEAD, the previously active branch will be reset to the
       target commit.
+    clone_opts : list of str, optional
+      Options passed to git-clone. Note that for RIA URLs, the version is
+      translated to a --branch argument, and that will take precedence over a
+      --branch argument included in this value.
 
     Yields
     ------
@@ -437,6 +457,7 @@ def clone_dataset(
         label='Clone attempt',
         unit=' Candidate locations',
     )
+    clone_opts = clone_opts or []
     error_msgs = OrderedDict()  # accumulate all error messages formatted per each url
     for cand in candidate_sources:
         log_progress(
@@ -446,17 +467,18 @@ def clone_dataset(
             update=1,
             increment=True)
 
-        clone_opts = {}
-
         if cand.get('version', None):
-            clone_opts['branch'] = cand['version']
+            opts = clone_opts + ["--branch=" + cand['version']]
+        else:
+            opts = clone_opts
+
         try:
             # TODO for now GitRepo.clone() cannot handle Path instances, and PY35
             # doesn't make it happen seamlessly
             GitRepo.clone(
                 path=str(dest_path),
                 url=cand['giturl'],
-                clone_options=clone_opts,
+                clone_options=opts,
                 create=True)
 
         except CommandError as e:
