### 2021-05-05 813a7a1
DOC: clone: Drop EnsureNone constraint from source parameter
The 'source' parameter of clone() includes EnsureNone in its
constraints, but downstream code doesn't handle a None value.
