### 2021-05-05 923198c
ENH: CLI - allow for no value for options with None available but not default
E.g. so we could have --fulfilled -- for foreach while having it to default to True
