### 2021-05-05 f2a4ca4
ENH: add {dspath} placeholder for consistency with run
