### 2021-05-05 99abdf6
RF: ProducerConsumer.get_effective_jobs to expose logic on deciding on number of jobs
