### 2021-05-05 5efed3f
ENH: add support for chpwd option defaulting to True (most common)
With safe-guard for python to go back to no threading if chpwd
