### 2021-05-05 12ff7bc
ENH: disable parallization for python mode and capture set for output streams
