### 2021-05-05 6a06a22
RF: switch to have pass-through to be default mode of handling output streams
I think this will be closer to a typical use case for use of foreach by a user:
just run a command and see what it spits out
