### 2021-05-05 31e0efc
ENH: get_status_dict - add exit_code if CommandError is passed
Immediate use-case is foreach since it feels worthwhile channeling
upstairs the details of the failed execution, but I can imagine
it being useful for other cases too
