### 2021-05-06 #3254 @csadorf
> 👋 @csadorf : would you be willing to review this article for JOSS?

Yes, I would be happy to! the only hitch is that I will be on vacation from May 13th until May 28th. What is the expected timeline?
