### 2021-05-07 #3254 @arokem
:wave: @pibion : would you be willing to review this article for JOSS?
