### 2021-05-07 #3254 @csadorf
> @csadorf : Ideally, we'd like to finish this in a few weeks. Do you think that you will have time for a first round of review before you head out on your vacation?

@arokem Since this would be my first review for JOSS I am afraid that I might not be able to without comprising due diligence. So unfortunately I have to decline this time. Please feel free to reach out to me again in case that you would still like my input on this particular submission after my return or for similar submissions in the future.
