### 2021-05-07 eb0aa98
[tmp] check appveyor stall
In the most recent PR (gh-5646), test_annexrepo.test_files_split looks
to be stalling:

  https://ci.appveyor.com/project/mih/datalad/builds/39042855/job/5hrhv3ueiqnc91r7

This scratch PR is to confirm that it occurs with master.  All of the
other builds are pruned.
