### 2021-05-07 434e3f9
ENH: customremotes: Add ensure_datalad_remote()
The datalad special remote is needed for a good amount of
functionality.  Add a helper that makes it easier for callers to
ensure that it's available.
