### 2021-05-07 1276e29
RF: download_url: Avoided repeated use of .repo attribute
