+++ datalad/interface/download_url.py
@@ -217,7 +217,8 @@ URLs:
                             on_failure="ignore"):
                 yield r
 
-            if isinstance(ds.repo, AnnexRepo):
+            ds_repo = ds.repo
+            if isinstance(ds_repo, AnnexRepo):
                 if got_ds_instance:
                     # Paths in `downloaded_paths` are already relative to the
                     # dataset.
@@ -238,7 +239,7 @@ URLs:
                                         orig_path, ds)
                 annex_paths = [p for p, annexed in
                                zip(rpaths,
-                                   ds.repo.is_under_annex(list(rpaths.keys())))
+                                   ds_repo.is_under_annex(list(rpaths.keys())))
                                if annexed]
                 if annex_paths:
                     for path in annex_paths:
@@ -246,7 +247,7 @@ URLs:
                         try:
                             # The file is already present. This is just to
                             # register the URL.
-                            ds.repo.add_url_to_file(
+                            ds_repo.add_url_to_file(
                                 path,
                                 url,
                                 # avoid batch mode for single files
@@ -261,4 +262,4 @@ URLs:
                     if archive:
                         from datalad.api import add_archive_content
                         for path in annex_paths:
-                            add_archive_content(path, annex=ds.repo, delete=True)
+                            add_archive_content(path, annex=ds_repo, delete=True)
