### 2021-05-07 ef6d1a5
BF: download-url: Set up datalad special remote if needed
download_url() downloads through DataLad's providers.  For URLs other
than unauthenticated ones, the URL won't be functional later if the
datalad special remote isn't enabled when the URL is registered with
git-annex.

If it looks like the datalad special remote is needed, handle the
setup automatically for the caller.

Closes #5510.
