### 2021-05-07 #3262 @yarikoptic
"Contributed equally" would be more appropriate indeed given current authors ordering.  Can/should we adjust manuscript and re-render here right away?
