### 2021-05-07 #3262 @yarikoptic
cool, so we don't forget I just committed/pushed to the [original repo](http://github.com/datalad/datalad-paper-joss) for now
