### 2021-05-10 fc70ace
Add a smoke test for #5631
_get_flexible_source_candidates_for_submodule should not crash with a KeyError if it can't match a template
