+++ CHANGELOG.md
@@ -9,13 +9,7 @@ This is a high level and scarce summary of the changes between releases.
 We would recommend to consult log of the 
 [DataLad git repository](http://github.com/datalad/datalad) for more details.
 
-## 0.14.4 (??? ??, 2021) -- will be better than ever
-
-bet we will fix some bugs and make a world even a better place.
-
-### Major refactoring and deprecations
-
-- hopefully none
+## 0.14.4 (May 10, 2021) -- .
 
 ### Fixes
 
@@ -26,10 +20,6 @@ bet we will fix some bugs and make a world even a better place.
 
 - Several test fixes, including updates for changes in git-annex.
   ([#5612][]) ([#5632][]) ([#5639][])
-
-### Enhancements and new features
-
-?
 
 
 ## 0.14.3 (April 28, 2021) -- .
