### 2021-05-10 cdc2b45
MNT: setup.py: Temporarily avoid Sphinx 4
With Sphinx 4.0.0 (released yesterday), the docs builds for maint and
master are failing (with different errors):

  https://github.com/datalad/datalad/actions/runs/828244083
  https://github.com/datalad/datalad/actions/runs/828227680

Hold off upgrading until the issues are resolved.
