### 2021-05-11 00ada4a
TST: download_url: Test datalad remote initialization
This test fails without the previous commit, which taught download-url
to initialize the datalad special remote if it is needed.

It requires credentials for s3://datalad-test0-versioned, so it will
not run in the CI builds.
