### 2021-05-11 4c303af
RF(TST): replace s3:// url for test with a publicly available
This is a file which is also used by h5py for testing of their support
of ROS3 driver.  Bucket is public and would not incure any cost.  File is tiny (4KB).

We could use a persistent VCR tape but it would have shaved off only about 2
seconds from total 14 seconds run time (hence adding @slow), and to make
it happen we would have needed to either copy helper code from
datalad/distribution/tests/test_create_github.py or just craft full path here.
So may be later if more usecases come up.

I have also harmonized imports a little while
[…]
