### 2021-05-11 ef73963
BF: create, gitrepo: Restore handling of initopts list
create() allows specifying git-init options via initopts, which can be
a dictionary (python callers) or list (command-line and python
callers).  In the list case, a dictionary of the form
{"_from_cmdline_": <list of options>} is constructed and passed to
GitRepo.

0c2dc0b596 (RF: Minimalist datalad.dataset.gitrepo interface,
2021-03-22) broke the handling of the list form by removing processing
of the "_from_cmdline_" special case:

  $ datalad create . --template=/dev/null
  [...]
  error: unknown option `-from-cmdline-=--template=/dev/nu
[…]
