### 2021-05-11 208e8d0
DOC: create: Correct documented type of initopts
_setup_git_repo() and_setup_annex_repo() take an initopts argument
that is destined for GitRepo.  The docstring of these functions say
that initopts is a list, but this isn't a form that GitRepo handles
properly.  It needs to be a dictionary, with list values snuck in
through the "_from_cmdline_" key.
