+++ datalad/core/local/create.py
@@ -462,7 +462,7 @@ def _setup_git_repo(path, initopts=None, fake_dates=False):
     ----------
     path: str or Path
       Path of the repository
-    initopts: list, optional
+    initopts: dict, optional
       Git options to be passed to the GitRepo constructor
     fake_dates: bool, optional
       Passed to the GitRepo constructor
@@ -502,7 +502,7 @@ def _setup_annex_repo(path, initopts=None, fake_dates=False,
     ----------
     path: str or Path
       Path of the repository
-    initopts: list, optional
+    initopts: dict, optional
       Git options to be passed to the AnnexRepo constructor
     fake_dates: bool, optional
       Passed to the AnnexRepo constructor
