### 2021-05-12 0cf73e4
DOC: Reference plugins via datalad.api
With the recent release of Sphinx v4, the doc build on master and
maint started to fail.  The failure on master was due to a bug in
Sphinx and fixed by v4.0.1.  The failure on maint is due to a warning:

    docstring of datalad.support.constraints.EnsureChoice:1:duplicate
    object description of datalad.support.constraints.EnsureChoice,
    other instance in generated/datalad.plugin.add_readme, use
    :noindex: for one of them

One our end, this is triggered by the combination of two things:

  * many of the plugin modules have imports at the class le
[…]
