### 2021-05-12 1548dfe
ENH: gitrepo.clone: Support plain list of options
GitRepo.clone() takes clone options as a dictionary (GitPython-style)
and converts them to a list of command line arguments.  Add support
for passing the list of arguments so that datalad-clone can directly
relay options from the command line.
