### 2021-05-12 29f0e75
RF: gitrepo.clone: Construct invariant command outside of loop
GitRepo.clone() calls git-clone within a loop so that it can retry the
same git-clone invocation to work around a known failure.  It defines
a "base command" outside of the loop and then extends that with the
path, url, and to_options(**clone_options) result within the loop.
This outside/within-loop split is potentially confusing to the reader
because the values that are tacked on per-iteration don't change.

Construct the entire command ahead of the loop.
