+++ datalad/support/gitrepo.py
@@ -993,19 +993,18 @@ class GitRepo(CoreGitRepo):
                     lgr.info("Expanded source path to %s from %s", new_url, url)
                     url = new_url
 
-        cmd_base = cls._git_cmd_prefix + ['clone', '--progress']
+        cmd = cls._git_cmd_prefix + ['clone', '--progress']
+        if clone_options:
+            cmd.extend(to_options(**clone_options))
+        cmd.extend([url, path])
+
         fix_annex = None
         ntries = 5  # 3 is not enough for robust workaround
         for trial in range(ntries):
             try:
                 lgr.debug("Git clone from {0} to {1}".format(url, path))
 
-                res = GitWitlessRunner().run(
-                        cmd_base + [url, path] \
-                        + (to_options(**clone_options)
-                           if clone_options else []),
-                        protocol=GitProgress,
-                )
+                res = GitWitlessRunner().run(cmd, protocol=GitProgress)
                 # fish out non-critical warnings by git-clone
                 # (empty repo clone, etc.), all other content is logged
                 # by the progress helper to 'debug'
