### 2021-05-13 c574032
ENH: install: Add --branch argument and relay to git-clone
The previous commit taught datalad-clone how to pass arbitrary
git-clone options to GitRepo.clone().  For datalad-install, we could
also allow arbitrary git-clone options.  However, given that
datalad-install wraps datalad-get in addition to datalad-clone, it's
not obvious that clone options should take precedence over arbitrary
git-annex-get options.

Instead add a dedicated --branch argument.  Callers that need the more
complicated options can use datalad-clone.  If it becomes clear that
datalad-install should support arbitrary clone
[…]
