### 2021-05-13 daa6285
RF: status: Avoid repeated Dataset.path access
