### 2021-05-13 3c68793
BF: status: Provide special treatment of "this dataset" path
When called with a path argument that points to a dataset and that
dataset has a parent dataset, status() checks if the original path
ends with "/".  No trailing separate indicates that the dataset as a
whole is of interest, leading to a query in the parent for that
dataset.

The above logic is confusing when `dataset` and `path` point to the
same dataset.  status() with a parent-less dataset works (and behaves
as the path argument ended with "/"), while a nested dataset fails
with

    [ERROR ] dataset containing given paths is not 
[…]
