+++ datalad/plugin/addurls.py
@@ -424,13 +424,39 @@ def _read_from_file(fname, input_type):
     return records, colidx_to_name
 
 
+_FIXED_SPECIAL_KEYS = {
+    "_repindex",
+    "_url_basename",
+    "_url_basename_ext",
+    "_url_basename_ext_py",
+    "_url_basename_root",
+    "_url_basename_root_py",
+    "_url_filename",
+    "_url_filename_ext",
+    "_url_filename_ext_py",
+    "_url_filename_root",
+    "_url_filename_root_py",
+    "_url_hostname",
+}
+
+
+def _is_known_special_key(key):
+    return key in _FIXED_SPECIAL_KEYS or re.match(r"\A_url[0-9]+\Z", key)
+
+
 def _get_placeholder_exception(exc, what, row):
     """Recast KeyError as a ValueError with close-match suggestions.
     """
     value = exc.args[0]
     if isinstance(value, str):
-        msg = "Unknown placeholder '{}' in {}: {}".format(
-            value, what, get_suggestions_msg(value, row))
+        if _is_known_special_key(value):
+            msg = ("Special key '{}' could not be constructed for row: {}"
+                   .format(value,
+                           {k: v for k, v in row.items()
+                            if not _is_known_special_key(k)}))
+        else:
+            msg = "Unknown placeholder '{}' in {}: {}".format(
+                value, what, get_suggestions_msg(value, row))
     else:
         msg = "Out-of-bounds or unsupported index {} in {}".format(
             value, what)
