### 2021-05-14 22a5b7a
RF: addurls._get_placeholder_exception: Rename a parameter
_get_placeholder_exception() callers pass in the row to provide a
collection of known keys, but an upcoming commit will care about
having access to the entire row.  Rename the parameter for clarity.
