### 2021-05-16 dc2b1aa
NF: implement a non asyncio-based runner
This patch implements a thread-based runner, that
uses the asyncio protocol interface to communicate
process output to protocol instances
