+++ datalad/cmd.py
@@ -50,70 +50,6 @@ _TEMP_std = sys.stdout, sys.stderr
 _MAGICAL_OUTPUT_MARKER = "_runneroutput_"
 
 
-async def run_async_cmd(loop, cmd, protocol, stdin, protocol_kwargs=None,
-                        **kwargs):
-    """Run a command in a subprocess managed by asyncio
-
-    This implementation has been inspired by
-    https://pymotw.com/3/asyncio/subprocesses.html
-
-    Parameters
-    ----------
-    loop : asyncio.AbstractEventLoop
-      asyncio event loop instance. Must support subprocesses on the
-      target platform.
-    cmd : list or str
-      Command to be executed, passed to `subprocess_exec` (list), or
-      `subprocess_shell` (str).
-    protocol : WitlessProtocol
-      Protocol class to be instantiated for managing communication
-      with the subprocess.
-    stdin : file-like or None
-      Passed to the subprocess as its standard input.
-    protocol_kwargs : dict, optional
-      Passed to the Protocol class constructor.
-    kwargs : Pass to `subprocess_exec`, will typically be parameters
-      supported by `subprocess.Popen`.
-
-    Returns
-    -------
-    undefined
-      The nature of the return value is determined by the given
-      protocol class.
-    """
-    if protocol_kwargs is None:
-        protocol_kwargs = {}
-    cmd_done = asyncio.Future(loop=loop)
-    factory = functools.partial(protocol, cmd_done, **protocol_kwargs)
-    kwargs.update(
-        stdin=stdin,
-        # ask the protocol which streams to capture
-        stdout=asyncio.subprocess.PIPE if protocol.proc_out else None,
-        stderr=asyncio.subprocess.PIPE if protocol.proc_err else None,
-    )
-    if isinstance(cmd, str):
-        proc = loop.subprocess_shell(factory, cmd, **kwargs)
-    else:
-        proc = loop.subprocess_exec(factory, *cmd, **kwargs)
-    transport = None
-    result = None
-    try:
-        lgr.debug('Launching process %s', cmd)
-        transport, protocol = await proc
-        lgr.debug('Waiting for process %i to complete', transport.get_pid())
-        # The next wait is a workaround that avoids losing the output of
-        # quickly exiting commands (https://bugs.python.org/issue41594).
-        await asyncio.ensure_future(transport._wait())
-        await cmd_done
-        result = protocol._prepare_result()
-    finally:
-        # protect against a crash when launching the process
-        if transport:
-            transport.close()
-
-    return result
-
-
 class WitlessProtocol(asyncio.SubprocessProtocol):
     """Subprocess communication protocol base class for `run_async_cmd`
 
@@ -128,29 +64,28 @@ class WitlessProtocol(asyncio.SubprocessProtocol):
     and a list with bytestrings of all captured output streams.
     """
 
-    FD_NAMES = ['stdin', 'stdout', 'stderr']
-
     proc_out = None
     proc_err = None
 
-    def __init__(self, done_future, encoding=None):
+    def __init__(self, encoding=None):
         """
         Parameters
         ----------
-        done_future : asyncio.Future
-          Future promise to be fulfilled when process exits.
         encoding : str
           Encoding to be used for process output bytes decoding. By default,
           the preferred system encoding is guessed.
         """
-        self.done = done_future
+
+        self.fd_infos = {}
+
+        self.process = None
+        self.stdout_fileno = 1
+        self.stderr_fileno = 2
+
         # capture output in bytearrays while the process is running
-        Streams = namedtuple('Streams', ['out', 'err'])
-        self.buffer = Streams(
-            out=bytearray() if self.proc_out else None,
[…]
