### 2021-05-17 d64e61d
RF: remove subprocess output logging
There are a number of places where datalad tests examine debug
output. Logging the output of sub processes throws those tests off.
