### 2021-05-17 862df37
Update datalad/nonasyncrunner.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
