+++ setup.py
@@ -7,8 +7,17 @@
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 
 from os.path import (
+    dirname,
     join as opj,
 )
+import sys
+
+# This is needed for versioneer to be importable when building with PEP 517.
+# See <https://github.com/warner/python-versioneer/issues/193> and links
+# therein for more information.
+sys.path.append(dirname(__file__))
+
+import versioneer
 
 from _datalad_build_support.setup import (
     BuildConfigInfo,
@@ -169,6 +178,9 @@ classifiers = [
     'Topic :: Utilities',
 ]
 setup_kwargs['classifiers'] = classifiers
+
+setup_kwargs["version"] = versioneer.get_version()
+cmdclass.update(versioneer.get_cmdclass())
 
 datalad_setup(
     'datalad',
