### 2021-05-17 e6b15bb
Update datalad/nonasyncrunner.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
