### 2021-05-17 ce51fa9
Update datalad/nonasyncrunner.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
