### 2021-05-17 2d526bd
BF: update: Honor --on-failure if fetch fails
update() calls fetch for each dataset it iterates over.  Catch a
CommandError and translate it to a result so that the caller can
decide how to handle the failure with --on-failure.

Closes #3278.
Closes #4154.
