### 2021-05-17 101613a
RF: update: Avoid fetch's deprecated kwargs
GitRepo.fetch's kwargs has been documented as deprecated since
v0.13.0.
