### 2021-05-18 b96bcd8
RF: add docstring, remove useless try-except-clause
BrokenPipeError is only raised on writing to a pipe,
we do not have to check for it here.
