### 2021-05-18 7bbccef
RF: return back datalad.version to not break "API" for maint
