### 2021-05-19 a9a51bd
RF: add tests, improve debug output
