+++ datalad/tests/test_nonasyncrunner.py
@@ -1,0 +1,125 @@
+# emacs: -*- mode: python-mode; py-indent-offset: 4; tab-width: 4; indent-tabs-mode: nil; coding: utf-8 -*-
+# ex: set sts=4 ts=4 sw=4 noet:
+# ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
+#
+#   See COPYING file distributed along with the datalad package for the
+#   copyright and license terms.
+#
+# ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
+"""Test the thread based runner (aka. non asyncio based runner).
+"""
+import logging
+import os
+import queue
+import subprocess
+import sys
+
+from datalad.tests.utils import assert_true, eq_
+
+from ..cmd import WitlessProtocol
+from ..nonasyncrunner import ReaderThread, run_command
+
+
+def test_subprocess_return_code_capture():
+
+    class KillProtocol(WitlessProtocol):
+
+        proc_out = True
+        proc_err = True
+
+        def __init__(self, signal_to_send: int, result_pool: dict):
+            super().__init__()
+            self.signal_to_send = signal_to_send
+            self.result_pool = result_pool
+
+        def connection_made(self, process):
+            super().connection_made(process)
+            process.send_signal(self.signal_to_send)
+
+        def connection_lost(self, exc):
+            self.result_pool["connection_lost_called"] = (True, exc)
+
+        def process_exited(self):
+            self.result_pool["process_exited_called"] = True
+
+    signal_to_send = 13
+    result_pool = dict()
+    result = run_command(["sleep", "10000"],
+                         KillProtocol,
+                         None,
+                         {
+                             "signal_to_send": signal_to_send,
+                             "result_pool": result_pool
+                         })
+    eq_(result["code"], -signal_to_send)
+    assert_true(result_pool["connection_lost_called"][0])
+    assert_true(result_pool["process_exited_called"])
+
+
+def test_interactive_communication():
+
+    class BidirectionalProtocol(WitlessProtocol):
+
+        proc_out = True
+        proc_err = True
+
+        def __init__(self, result_pool: dict):
+            super().__init__()
+            self.state = 0
+            self.result_pool = result_pool
+
+        def connection_made(self, process):
+            super().connection_made(process)
+            os.write(self.process.stdin.fileno(), b"1 + 1\n")
+
+        def connection_lost(self, exc):
+            self.result_pool["connection_lost_called"] = True
+
+        def process_exited(self):
+            self.result_pool["process_exited_called"] = True
+
+        def pipe_data_received(self, fd, data):
+            super().pipe_data_received(fd, data)
+            if self.state == 0:
+                self.state += 1
+                os.write(self.process.stdin.fileno(), b"2 ** 3\n")
+            if self.state == 1:
+                self.state += 1
+                os.write(self.process.stdin.fileno(), b"exit(0)\n")
+
+    result_pool = dict()
+    result = run_command(["python3", "-i"],
+                         BidirectionalProtocol,
+                         stdin=subprocess.PIPE,
+                         protocol_kwargs={
+                            "result_pool": result_pool
+                         })
+
+    eq_(result["stdout"], "2\n8\n")
+    assert_true(result_pool["connection_lost_called"], True)
+    assert_true(result_pool["process_exited_called"], True)
+
+
+def test_thread_exit():
+    logger = logging.getLogger("datalad.runner")
+    logger.setLevel(logging.DEBUG)
+
+    (read_descriptor, write_descriptor) = os.pipe()
+    read_file = os.fdopen(read_descriptor, "r")
[…]
