### 2021-05-19 179a2cf
NF: add test_popen_invocation
