### 2021-05-19 14a7f0c
BF: handle different line endings in nonasync tests
