### 2021-05-19 a33da5d
upgrade with devel version of versioneer to get sensible version from shallow clones
had to manually reset __init__.py in current approach, see
https://github.com/python-versioneer/python-versioneer/issues/247
