### 2021-05-19 e2330fc
TST: Add ORA reconfiguration test
Test, that a configured push-url for ORA remote that matches the URL we
clone from, does prevent `clone` from reconfiguring the ORA remote. Just
like a matching `url` would.
See gh-5628 for usage pattern and implications.
