### 2021-05-19 005b5f1
Update datalad/cmd.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
