### 2021-05-19 bd7bb59
Update datalad/cmd.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
