### 2021-05-19 03b2ced
ENH: addurls: Report number of colliding file names in error message
When there are file name collisions, more details are logged at the
debug level, but rephrase the error message to give the caller a sense
of how many conflicts there are.
