### 2021-05-19 8325c5b
RF: addurls: Store collisions as intermediate result
addurls() checks if there a collision by taking the set() of file
names across rows and checking that the length is equal to the number
of total rows.  In preparation for letting the caller specify
different ways to handle the collision (e.g., take the last row),
rework the logic to store the collisions as a dictionary.
