### 2021-05-19 48db24b
ENH: addurls: Provide more logging output for debugging collisions
When more than one row produces a file name, the error message just
reports that there are collisions, leaving the caller to figure out
where.

Help the caller by providing 1) the row positions that conflict for
each file name and 2) a sample of two rows that conflict.  Given that
there may be many conflicts and that an input row may be very long,
the error message of the result doesn't seem like a good place to put
this information.  Instead log it at the debug level.
