### 2021-05-19 ef61f31
ENH: addurls: Support taking first/last row when file names collide
The previous commit taught addurls() to ignore file name collisions
when the colliding rows have the same URL and metadata.  Go one step
farther and allow the caller to say "I don't care if they're
different, just take the {first,last} row and ignore the rest".
