+++ datalad/local/addurls.py
@@ -598,12 +598,61 @@ def _find_collisions(rows):
     return {fname: fname_idxs[fname] for fname in collisions}
 
 
-def _handle_collisions(rows):
-    """Handle file name collisions in `rows`.
+def _find_collision_mismatches(rows, collisions):
+    """Find collisions where URL and metadata fields don't match.
 
     Parameters
     ----------
     rows : list of dict
+    collisions : dict
+        File names with collisions mapped to positions in `rows` that have a
+        given file name.
+
+    Returns
+    -------
+    Dict with subset of `collisions` where at least one colliding row has a
+    different URL or metadata field value.
+    """
+    def get_key(row):
+        return row["url"], row.get("meta_args")
+
+    mismathches = {}
+    for fname, idxs in collisions.items():
+        key_0 = get_key(rows[idxs[0]])
+        if any(key_0 != get_key(rows[i]) for i in idxs[1:]):
+            mismathches[fname] = idxs
+    return mismathches
+
+
+def _ignore_collisions(rows, collisions):
+    """Modify `rows`, marking those that produce collision as ignored.
+
+    Parameters
+    ----------
+    rows : list of dict
+    collisions : dict
+        File names with collisions mapped to positions in `rows` that have a
+        given file name.
+    """
+    for fname in collisions:
+        for idx in collisions[fname][:-1]:
+            lgr.debug("Ignoring collision of file name '%s' at row %d",
+                      fname, rows[idx]["input_idx"])
+            rows[idx]["ignore"] = True
+
+
+def _handle_collisions(rows, on_collision):
+    """Handle file name collisions in `rows`.
+
+    "Handling" consists of either marking all but one colliding row with
+    ignore=True or returning an error message, depending on the value of
+    `on_collision`. When an error message is returned, downstream processing of
+    the rows should not be done.
+
+    Parameters
+    ----------
+    rows : list of dict
+    on_collision : {"error", "error-if-different"}
 
     Returns
     -------
@@ -612,8 +661,19 @@ def _handle_collisions(rows):
     err_msg = None
     collisions = _find_collisions(rows)
     if collisions:
-        err_msg = ("There are file name collisions; "
-                   "consider using {_repindex}")
+        if on_collision == "error":
+            to_report = collisions
+        elif on_collision == "error-if-different":
+            to_report = _find_collision_mismatches(rows, collisions)
+        else:
+            raise ValueError(
+                f"Unsupported `on_collision` value: {on_collision}")
+
+        if to_report:
+            err_msg = ("There are file name collisions; "
+                       "consider using {_repindex}")
+        else:
+            _ignore_collisions(rows, collisions)
     return err_msg
 
 
@@ -1184,6 +1244,13 @@ class Addurls(Interface):
             action="store_true",
             doc="""drop files after adding to annex""",
         ),
+        on_collision=Parameter(
+            args=("--on-collision",),
+            constraints=EnsureChoice("error", "error-if-different"),
+            doc="""What to do when more than one row produces the same file
+            name. By default an error is triggered. "error-if-different"
+            suppresses that error if rows for a given file name collision have
+            the same URL and metadata."""),
     )
 
     @staticmethod
@@ -1193,7 +1260,8 @@ class Addurls(Interface):
                  input_type="ext", exclude_autometa=None, meta=None, key=None,
                  message=None, dry_run=False, fast=False, ifexists=None,
                  missing_value=None, save=True, version_urls=False,
-                 cfg_proc=None, jobs=None, drop_after=False):
+                 cfg_proc=None, jobs=None, drop_after=False,
+                 on_collision="error"):
         # This was to work around gh-2269. That's fixed, but changing the
         # positional argument names now would cause breakage for any callers
[…]
