### 2021-05-19 b0b1882
RF: addurls: Add dedicated function for handling collisions
The logic around collisions is going to get more involved.  Move it to
a helper to avoid blowing up an already-too-long __call__().
