### 2021-05-19 a841bc0
RF/BF: do not make clone shallow in appveyor
