### 2021-05-19 70b1351
BF: fix result construction, improve log messages
