+++ datalad/core/distributed/tests/test_clone.py
@@ -629,7 +629,7 @@ def test_cfg_originorigin(path):
     with chpwd(path), swallow_logs(new_level=logging.DEBUG) as cml:
         clone_lev3 = clone('clone_lev2', 'clone_lev3')
         # we called git-annex-init; see gh-4367:
-        cml.assert_logged(msg=r"[^[]*Async run:\n cwd=.*\n"
+        cml.assert_logged(msg=r"[^[]*run:\n cwd=.*\n"
                               r" cmd=\[('git',.*'annex'|'git-annex'), 'init'",
                           match=False,
                           level='DEBUG')
