### 2021-05-19 d1888cf
RF: add type hints, improve doc-strings, improve tests
