### 2021-05-19 80b183a
RF: just pip install datalad-installer
