+++ .github/workflows/test_macos.yml
@@ -24,8 +24,8 @@ jobs:
 
     - name: Install git-annex
       run: |
-        wget -O datalad_installer.py https://raw.githubusercontent.com/datalad/datalad-installer/master/src/datalad_installer.py
-        python3 datalad_installer.py --sudo ok -E new.env git-annex -m ${{ matrix.install_scenario }}
+        pip install datalad-installer
+        datalad-installer --sudo ok -E new.env git-annex -m ${{ matrix.install_scenario }}
         . new.env
         echo "PATH=$PATH" >> "$GITHUB_ENV"
 
