### 2021-05-20 e27c6bf
RF: Establish common keyring get helper in Credential
