### 2021-05-20 ec0b638
Relax permission requirements for RIA stores
Previously the ORA special remote relied on whatever the store end
is set up to with respect to permissions and accepted them as is,
meaning it didn't try to obtain write permissions on the parent
directory when dropping a key from store. However, there are good
reasons to revoke write permissions in store, particularly when the
object tree is meant to be used with ephemeral clones.

Hence, do what git-annex would do in the local case: View that as
protection against direct user interaction and try to obtain permission
on drop.

(Closes #6254)
