### 2021-05-20 78e85f9
TST: Add a test for issue 5668
Test whether the ORA remote can deal with an object tree in a RIA store,
where key directories have no write permissions (as annex would do).
