### 2021-05-20 f75c731
NF: Support credential query via ConfigManager
Fixes gh-5677
