### 2021-05-20 7f6cc85
RF: improve test coverage
