### 2021-05-20 71e1a26
NF: Config setting datalad.ui.suppress-similar-results
That can be used to determine the default result suppression
setup, or for ad-hoc switching.

Closes gh-4417
