+++ datalad/interface/utils.py
@@ -558,7 +558,10 @@ def _process_results(
     # counter for detected repetitions
     result_repetitions = 0
     # how many repetitions to show, before suppression kicks in
-    render_n_repetitions = 10 if sys.stdout.isatty() else float("inf")
+    render_n_repetitions = 10 \
+        if sys.stdout.isatty() \
+        and dlcfg.obtain('datalad.ui.suppress-similar-results') \
+        else float("inf")
 
     for res in results:
         if not res or 'action' not in res:
