### 2021-05-20 311f588
NF: Config setting datalad.ui.suppress-similar-results-threshold
To make result suppression threshold configurable.
