### 2021-05-20 144f8f9
RF: improve test coverage of WitlessProtocol
