### 2021-05-20 e8e43ff
RF: remove unused code from WitlessRunner
