### 2021-05-20 d9bffb9
Install python 3.6 first and then installer (needs python3)
