+++ .github/workflows/test_macos.yml
@@ -22,6 +22,11 @@ jobs:
 
     - uses: actions/checkout@v1
 
+    - name: Set up Python 3.6
+      uses: actions/setup-python@v1
+      with:
+        python-version: 3.6
+    
     - name: Install git-annex
       run: |
         pip install datalad-installer
@@ -29,10 +34,6 @@ jobs:
         . new.env
         echo "PATH=$PATH" >> "$GITHUB_ENV"
 
-    - name: Set up Python 3.6
-      uses: actions/setup-python@v1
-      with:
-        python-version: 3.6
 
     - name: Install dependencies
       run: |
