### 2021-05-20 854ca94
BF+ENH: use python -m pip to not fall into python2 + allow to override installer version
