### 2021-05-20 datalad/datalad#5685 @jkanche
Add LICENSE to the repo
I notice the license in the `README`, but would be great to explicitly add a `LICENSE` file to the repo
