### 2021-05-20 04f21c3
ENH/UX: Record the use of --explicit in the run record of a commit
