### 2021-05-20 33a3e5f
ENH: Detect whether a part run used --explicit, and set it too when the commit is rerun
BF: Fix erroneous cascade of 'explicit=True' configurations

Thanks to Kyle for the patch. Without it, any explicit=True would be propagated into subsequent commits,
irrespective of whether their commit included an --explicit flag
