### 2021-05-20 datalad/datalad#5685 @kyleam
It's there, just with a different name: COPYING.
