### 2021-05-20 datalad/datalad#5685 @adswa
Hey @jkanche, I think this comes in the context of the JOSS review, right? As @kyleam said, we keep the license (MIT) in the ``COPYING`` file - how strict do you think is that item on the review checklist to be interpreted? :smile: I actually don't know where difference in naming comes from, but judging from https://stackoverflow.com/questions/5678462/should-i-provide-a-license-txt-or-copying-txt-file-in-my-project  both are in standard use.
