+++ datalad/cmd.py
@@ -11,17 +11,24 @@ Wrapper for command and function calls, allowing for dry runs and output handlin
 
 """
 
-import subprocess
-import sys
 import logging
 import os
+import subprocess
+import sys
 import tempfile
-from locale import getpreferredencoding
-import asyncio
 import warnings
 
+from .cmd_protocols import (
+    KillOutput,
+    NoCapture,
+    StdErrCapture,
+    StdOutCapture,
+    StdOutErrCapture,
+    WitlessProtocol,
+)
 from .consts import GIT_SSH_COMMAND
 from .dochelpers import borrowdoc
+from .nonasyncrunner import run_command
 from .support import path as op
 from .support.exceptions import CommandError
 from .utils import (
@@ -42,161 +49,6 @@ _TEMP_std = sys.stdout, sys.stderr
 # in Runner so we take care about their removal, in contrast to those
 # which might be created outside and passed into Runner
 _MAGICAL_OUTPUT_MARKER = "_runneroutput_"
-
-
-class WitlessProtocol(asyncio.SubprocessProtocol):
-    """Subprocess communication protocol base class for `run_async_cmd`
-
-    This class implements basic subprocess output handling. Derived classes
-    like `StdOutCapture` should be used for subprocess communication that need
-    to capture and return output. In particular, the `pipe_data_received()`
-    method can be overwritten to implement "online" processing of process
-    output.
-
-    This class defines a default return value setup that causes
-    `run_async_cmd()` to return a 2-tuple with the subprocess's exit code
-    and a list with bytestrings of all captured output streams.
-    """
-
-    proc_out = None
-    proc_err = None
-
-    def __init__(self, done_future=None, encoding=None):
-        """
-        Parameters
-        ----------
-        done_future: Any
-          Ignored parameter, kept for backward compatibility (DEPRECATED)
-        encoding : str
-          Encoding to be used for process output bytes decoding. By default,
-          the preferred system encoding is guessed.
-        """
-
-        if done_future is not None:
-            warnings.warn("`done_future` argument is ignored "
-                          "and will be removed in a future release",
-                          DeprecationWarning)
-
-        self.fd_infos = {}
-
-        self.process = None
-        self.stdout_fileno = 1
-        self.stderr_fileno = 2
-
-        # capture output in bytearrays while the process is running
-        self.fd_infos[self.stdout_fileno] = ("stdout", bytearray()) if self.proc_out else ("stdout", None)
-        self.fd_infos[self.stderr_fileno] = ("stderr", bytearray()) if self.proc_err else ("stderr", None)
-
-        super().__init__()
-        self.encoding = encoding or getpreferredencoding(do_setlocale=False)
-
-        self._log_outputs = False
-        if lgr.isEnabledFor(5):
-            try:
-                from . import cfg
-                self._log_outputs = cfg.getbool('datalad.log', 'outputs', default=False)
-            except ImportError:
-                pass
-            self._log = self._log_summary
-        else:
-            self._log = self._log_nolog
-
-    def _log_nolog(self, *args):
-        pass
-
-    def _log_summary(self, fd, data):
-        fd_name = self.fd_infos[fd]
-        lgr.log(5, 'Read %i bytes from %i[%s]%s',
-                len(data), self.process.pid, fd_name, ':' if self._log_outputs else '')
-        if self._log_outputs:
-            log_data = ensure_unicode(data)
-            # The way we log is to stay consistent with Runner.
-            # TODO: later we might just log in a single entry, without
-            # fd_name prefix
-            lgr.log(5, "%s| %s ", fd_name, log_data)
-
-    def connection_made(self, process):
-        self.process = process
-        lgr.debug('Process %i started', self.process.pid)
-
-    def pipe_data_received(self, fd, data):
-        self._log(fd, data)
[…]
