### 2021-05-20 c13cfb5
ENH: addurls: Improve result displayed for subdataset creation
When addurls() creates a subdataset, an unhelpful

  create(ok): . (dataset)

is displayed.  That's because the result comes directly from
<subds>.create().

To display the subdataset path, one option would be to intercept the
result and adjust the refds field to point to the top-level addurls()
dataset.  However, this is a problematic change because it would break
any result hooks.  Aside from that, it's an artificial result that
could mislead and confuse those inspecting it.

Instead define a custom result renderer that limits th
[…]
