### 2021-05-20 eca3477
RF: interface: Move default summary logic to dedicated function
As mentioned in the last commit, there are cases where a custom
summary renderer wants to generate the default summary display.  As of
the last commit, they can avoid recalculating the summaries by
instructing eval_results() to pass the action summary.

The last piece needed to be able to easily generate the default
summary is the formatting logic for the action summary.  Extract that
to a dedicated function so that custom_result_summary_renderer()
definitions can use it.
