### 2021-05-20 633119c
MNT: interface.tests.test_base: Simplify test setup
A comment in test_status_custom_summary_no_repeats() mentions that the
test could be simplified once combined with another in-flight commit.
That commit is now merged.

I've confirmed that this still serves as a regression test (i.e. it
fails if 99124fc15d is reverted).
