### 2021-05-20 8e8ae6a
BF: interface: Don't repeat custom summary if non-generator results
99124fc15d (BF: interface: Drop repeated
custom_result_summary_renderer call, 2020-04-30) fixed a case where
the custom summary is duplicated.  The analysis there mentions that at
the time there were three custom_result_summary_renderer() call sites.
That commit dropped one of the calls to solve the duplicated output on
the command line.

However, it's still possible to trigger _both_ of the remaining calls
from the Python interface, leading to duplicate output.  The two
remaining call sites are

  * in the return_type!="gener
[…]
