### 2021-05-20 994ad6d
TST: addurls: Call command with result_renderer=None
The next commit will change addurls() to use a tailored result
renderer.  As a consequence, calls to addurls() will not use
result_renderer=None by default.

Explicitly specify result_renderer=None to maintain the current output
noise level.
