+++ datalad/local/tests/test_addurls.py
@@ -350,7 +350,8 @@ def test_extract_wrong_input_type():
 def test_addurls_nonannex_repo(path):
     ds = Dataset(path).create(force=True, annex=False)
     with assert_raises(IncompleteResultsError) as raised:
-        ds.addurls("dummy_arg0", "dummy_arg1", "dummy_arg2")
+        ds.addurls("dummy_arg0", "dummy_arg1", "dummy_arg2",
+                   result_renderer=None)
     assert_in("not an annex repo", str(raised.exception))
 
 
@@ -359,26 +360,31 @@ def test_addurls_unknown_placeholder(path):
     ds = Dataset(path).create(force=True)
     # Close but wrong URL placeholder
     with assert_raises(IncompleteResultsError) as exc:
-        ds.addurls("in.csv", "{link}", "{abcd}", dry_run=True)
+        ds.addurls("in.csv", "{link}", "{abcd}", dry_run=True,
+                   result_renderer=None)
     assert_in("linky", str(exc.exception))
     # Close but wrong file name placeholder
     with assert_raises(IncompleteResultsError) as exc:
-        ds.addurls("in.csv", "{linky}", "{abc}", dry_run=True)
+        ds.addurls("in.csv", "{linky}", "{abc}", dry_run=True,
+                   result_renderer=None)
     assert_in("abcd", str(exc.exception))
     # Out-of-bounds index.
     with assert_raises(IncompleteResultsError) as exc:
-        ds.addurls("in.csv", "{linky}", "{3}", dry_run=True)
+        ds.addurls("in.csv", "{linky}", "{3}", dry_run=True,
+                   result_renderer=None)
     assert_in("index", str(exc.exception))
 
     # Suggestions also work for automatic file name placeholders
     with assert_raises(IncompleteResultsError) as exc:
-        ds.addurls("in.csv", "{linky}", "{_url_hostnam}", dry_run=True)
+        ds.addurls("in.csv", "{linky}", "{_url_hostnam}", dry_run=True,
+                   result_renderer=None)
     assert_in("_url_hostname", str(exc.exception))
     # ... though if you whiff on the beginning prefix, we don't suggest
     # anything because we decide to generate those fields based on detecting
     # the prefix.
     with assert_raises(IncompleteResultsError) as exc:
-        ds.addurls("in.csv", "{linky}", "{_uurl_hostnam}", dry_run=True)
+        ds.addurls("in.csv", "{linky}", "{_uurl_hostnam}", dry_run=True,
+                   result_renderer=None)
     assert_not_in("_url_hostname", str(exc.exception))
 
 
@@ -399,7 +405,7 @@ def test_addurls_dry_run(path):
         ds.addurls(json_file,
                    "{url}",
                    "{subdir}//{_url_filename_root}",
-                   dry_run=True)
+                   dry_run=True, result_renderer=None)
 
         for dir_ in ["foo", "bar"]:
             assert_in("Would create a subdataset at {}".format(dir_),
@@ -477,7 +483,7 @@ class TestAddurls(object):
         json_file = op.relpath(self.json_file, ds.path)
 
         ds.addurls(json_file, "{url}", "{name}",
-                   exclude_autometa="(md5sum|size)")
+                   exclude_autometa="(md5sum|size)", result_renderer=None)
         ok_startswith(ds.repo.format_commit('%b', DEFAULT_BRANCH), f"url_file='{json_file}'")
 
         filenames = ["a", "b", "c"]
@@ -502,19 +508,20 @@ class TestAddurls(object):
         # Add to already existing links, overwriting.
         with swallow_logs(new_level=logging.DEBUG) as cml:
             ds.addurls(self.json_file, "{url}", "{name}",
-                       ifexists="overwrite")
+                       ifexists="overwrite", result_renderer=None)
             for fname in filenames:
[…]
