### 2021-05-21 4dbaecb
TST: interface: Test custom summary functionality
push() and status() define a custom summary renderer, but there are no
dedicated tests for the custom_result_summary_renderer() interface.
Add one.
