### 2021-05-21 20de96c
ENH: interface: Expose action summary to custom summary methods
When processing results, eval_results() builds up an action summary
dictionary that tracks that status counts for each command.  This
information is used to generate the default result summary.

Commands may also define a custom summary method.  This gets passed a
list of all results, which it can summarize however it wants.
However, this is a bit wasteful in two use cases:

  * the method wants to include information _in addition_ to the
    standard action summary

  * if a command uses a custom renderer, the default summary isn
[…]
