### 2021-05-21 02fa519
add --on-failure default value and document it
