### 2021-05-21 36aee8b
Just take a version of master altogether (will need fixups)
the only not applied change is removal of datalad.plugin testing -- still
in maint
