### 2021-05-21 0e759af
Strengthen apt-get with an update. Added a comment on alternative git-annex installation way
