### 2021-05-21 c628af3
RF: disable override of INSTALL_GITANNEX (had none in maint)
