### 2021-05-21 ec0243c
BF: get_content_{annex,}info: Restrict full report to paths=None
GitRepo.get_content_info() and AnnexRepo.get_content_annexinfo()
report on all paths in a tree if the `paths` argument is false-y,
including if it is an empty list.

This behavior requires callers to guard against the empty list case if
they really want to restrict results to the paths in the list.  That
means it's usually not safe to do something like

  get_content_info(paths=[<some filtering list comprehension>], ...)

which was the source of the bug in gh-3288.

This behavior is also problematic because GitRepo.diff()
GitRepo
[…]
