### 2021-05-22 e63b17d
RF: get_content_annexinfo: Mark content availability just once
get_content_annexinfo() calls _mark_content_availability() for each
file received from `git annex find{,ref}`.  This is unnecessary work
because _mark_content_availability() operates on the entire dictionary
of paths.  Delay the call until after get_content_annexinfo() has
constructed the entire dictionary from the `git annex find{,ref}`
information.
