### 2021-05-24 d524cd9
TST: test_fileinfo: Be more specific about status() return value
Re: https://github.com/datalad/datalad/pull/5693#pullrequestreview-666722514
