### 2021-05-25 a5d554b
ENH: run: Provide result record for command execution
After executing a command, yield a result record.  This allows callers
to control whether the save happens via the standard --on-failure
switch.

Note that this contains two breaking changes:

  * it changes a CommandError to a result record with status="error"

  * the on_failure is still "continue", so run() now saves failed
    commands by default

Suggested by @mih in gh-3071.
