### 2021-05-25 60876b4
ENH: interface: Consistently expose common parameters to renderers
eval_results() combines the positional and keyword arguments of a
command's __call__() into an OrderedDict.  This has been passed to
result filters since 141a55f9c0 (RF: Make positionals also available
as kwargs to result_filter, 2017-04-24) and to custom result renderers
since 6267129163 (BF: Result renderers never got to see all args of a
command call, 2019-12-23).

Whether the dictionary of arguments contains eval_params parameters
depends on how the command is called.  From the command line,
return_type, result_renderer, an
[…]
