+++ datalad/interface/utils.py
@@ -271,11 +271,6 @@ def eval_results(wrapped):
     @wraps(wrapped)
     def eval_func(*args, **kwargs):
         lgr.log(2, "Entered eval_func for %s", wrapped)
-        # for result filters
-        # we need to produce a dict with argname/argvalue pairs for all args
-        # incl. defaults and args given as positionals
-        allkwargs = get_allargs_as_kwargs(wrapped, args, kwargs)
-
         # determine the command class associated with `wrapped`
         wrapped_class = get_wrapped_class(wrapped)
 
@@ -290,6 +285,12 @@ def eval_results(wrapped):
                 # default set in that class
                 getattr(wrapped_class, p_name))
             for p_name in eval_params}
+
+        # for result filters
+        # we need to produce a dict with argname/argvalue pairs for all args
+        # incl. defaults and args given as positionals
+        allkwargs = get_allargs_as_kwargs(wrapped, args,
+                                          {**kwargs, **common_params})
 
         # short cuts and configured setup for common options
         return_type = common_params['return_type']
