### 2021-05-25 c48dc4e
[maybe] ENH: run: Change on_failure default to "stop"
As of the last commit, run() now saves failed commands by default.
Make run() use on_failure=stop by default to restore the old behavior.

This of course also changes the default behavior for upstream records,
in particular get() and unlock() ones.
