### 2021-05-25 9595f5e
ENH: run: Provide result record for command execution
After executing a command, yield a result record.  This allows callers
to control whether the save happens via the standard --on-failure
switch.

Note that this contains a breaking change:

  * it changes a CommandError to a result record with status="error"

Suggested by @mih in gh-3071.
