+++ datalad/core/local/run.py
@@ -318,6 +318,23 @@ class Run(Interface):
             else:
                 raise ValueError(f"Unknown dry-run mode: {dry_run!r}")
         else:
+            if kwargs.get("on_failure") == "stop" and \
+               res.get("action") == "run" and res.get("status") == "error":
+                msg_path = res.get("msg_path")
+                if msg_path:
+                    if datalad.get_apimode() == 'python':
+                        ds_path = res["path"]
+                        help = f"\"Dataset('{ds_path}').save(path='.', " \
+                               "recursive=True, message_file='%s')\""
+                    else:
+                        help = "'datalad save -d . -r -F %s'"
+                    lgr.info(
+                        "The command had a non-zero exit code. "
+                        "If this is expected, you can save the changes with "
+                        f"{help}",
+                        # shorten to the relative path for a more concise
+                        # message
+                        msg_path.relative_to(ds.pathobj))
             generic_result_renderer(res)
 
 
@@ -580,7 +597,7 @@ def format_command(dset, command, **kwds):
     return sfmt.format(command, **kwds)
 
 
-def _execute_command(command, pwd, expected_exit=None):
+def _execute_command(command, pwd):
     from datalad.cmd import WitlessRunner
 
     exc = None
@@ -595,16 +612,6 @@ def _execute_command(command, pwd, expected_exit=None):
     except CommandError as e:
         exc = e
         cmd_exitcode = e.code
-
-        if expected_exit is not None and expected_exit != cmd_exitcode:
-            # we failed in a different way during a rerun.  This can easily
-            # happen if we try to alter a locked file
-            #
-            # TODO add the ability to `git reset --hard` the dataset tree on failure
-            # we know that we started clean, so we could easily go back, needs gh-1424
-            # to be able to do it recursively
-            raise exc
-
     lgr.info("== Command exit (modification check follows) =====")
     return cmd_exitcode or 0, exc
 
@@ -774,9 +781,7 @@ def run_command(cmd, dataset=None, inputs=None, outputs=None, expand=None,
         return
 
     if not inject:
-        cmd_exitcode, exc = _execute_command(
-            cmd_expanded, pwd,
-            expected_exit=rerun_info.get("exit", 0) if rerun_info else None)
+        cmd_exitcode, exc = _execute_command(cmd_expanded, pwd)
         run_info['exit'] = cmd_exitcode
 
     # Re-glob to capture any new outputs.
@@ -823,30 +828,34 @@ def run_command(cmd, dataset=None, inputs=None, outputs=None, expand=None,
     if outputs_to_save is not None and use_sidecar:
         outputs_to_save.append(record_path)
     do_save = outputs_to_save is None or outputs_to_save
+    msg_path = None
     if not rerun_info and cmd_exitcode:
         if do_save:
             repo = ds.repo
-            msg_path = repo.dot_git / "COMMIT_EDITMSG"
-            msg_path.write_text(msg)
-            if datalad.get_apimode() == 'python':
-                help = f"\"Dataset('{ds.path}').save(path='.', " \
-                       "recursive=True, message_file='%s')\""
-            else:
-                help = "'datalad save -d . -r -F %s'"
-            lgr.info("The command had a non-zero exit code. "
-                     "If this is expected, you can save the changes with "
-                     f"{help}",
-                     # shorten to the relative path for a more concise message
-                     msg_path.relative_to(ds.pathobj))
-            if repo.dirty and not explicit:
-                # Give clean-up hints if a formerly clean repo is left dirty
-                lgr.info(
-                    "The commands 'git clean -di' and/or 'git reset' "
-                    "could be used to get to a clean dataset state again. "
[…]
