### 2021-05-25 e4d045e
CLN: run: Delete stale comment
When run() was first added, it used to take the exit code from the
CommandError raised by a failed command and create a fresh
CommandError with the exit code.  That hasn't been the case since
3817fb4f81 (ENH: LEt `run` spit out the original full-detail error,
2018-04-25).
