### 2021-05-25 2b5b29a
ENH: push: Tack action summary onto custom summary renderer
push() has a custom summary renderer that collects and renders hints
from git-push, but it's still useful to see the standard action count
summary.

Closes #5660.
