### 2021-05-26 6b9e811
TST: push: Add a basic test of custom summary renderer
Note that I've left "# pragma: more cover" on
Push.custom_result_summary_renderer because some of the logic (e.g.,
getting unique hints) still isn't tested.
