### 2021-05-26 38fe6b1
DOC: get_allargs_as_kwargs: Fix a docstring typo
