### 2021-05-26 cb19377
TST: Drop skip_v6_or_later and known_failure_v6_or_later
With the new minimum git-annex version, v5 auto-upgrades to v8, so
everything is v6+.

Neither of these appear to be used in the latest release of any
extension.
