### 2021-05-26 9cae34e
BF: common_cfg: Update default for datalad.repo.version
Using 5 for the default value of datalad.repo.version hasn't been a
good value since v5 git-annex started auto-upgrading to v7+
(7.20190912), and it should have certainly been updated once DataLad's
minimum git-annex version was one that only supported v8 [*].

Note though that having a stale value here is harmless because 1) the
default is not actually used by the code that feeds the value to
git-annex-init and 2) even if it were, git-annex-init would just
auto-upgrade it.

[*] 315911ca76 (MNT: annexrepo: Bump minimum git-annex version t
[…]
