### 2021-05-26 50d410d
TST: test_run_procedure: Prune annex v5 condition
With the new minimum git-annex version, v5 auto-upgrades to v8, so
everything is v6+.
