### 2021-05-26 71c7c84
ENH: annexrepo: Inform users about repo version auto-upgrades
The current minimum version of git-annex (8.20200309) auto-upgrades
v3-7 to v8.  Let the user know that, despite having explicitly
specified a version, they will end up with a different one.

We tried this once before in abed6efad8 (ENH: annexrepo: Inform users
about repo version auto-upgrades, 2018-11-29), but that was reverted
due to triggering a stall in test_annex_ssh().  Since then, git-annex
has had a good number of SSH-related fixups and test_annex_ssh() has
also undergone major changes, so let's hope that problem has gone
aw
[…]
