### 2021-05-26 eb5cdae
RF: annexrepo: Convert datalad.repo.version to an int
AnnexRepo.__init__() uses config.get() to get datalad.repo.version,
which will return the version as a string.  obtain() could be used
instead and would handle the conversion to int, but that doesn't work
well here because it's better for git-annex-init to determine
