### 2021-05-31 6aa1d71
BF(TST): skip testing for having "scanning for unlocked" with recent git-annex
Closes #5702

Recently, somewhere around [8.20210428-186-g428c91606](https://git.kitenet.net/index.cgi/git-annex.git/commit/?id=428c91606b434512d1986622e751c795edf4df44)
git-annex changed its logic for doing initial scanning for
unlocked (and now locked?) files.  Now I do not observe that message
to be printed out and our tests started to fail in daily testing of
datalad/git-annex
https://github.com/datalad/git-annex/actions/runs/877176725 might be
the first one and version reported is 8.20210428+git228-g13a6bfff4
w
[…]
