+++ datalad/support/tests/test_annexrepo.py
@@ -1306,7 +1306,11 @@ def test_init_scanning_message(path):
     # | end kludge
     with swallow_logs(new_level=logging.INFO) as cml:
         AnnexRepo(path, create=True, version=7)
-        assert_in("for unlocked", cml.out)
+        # somewhere around 8.20210428-186-g428c91606 git annex changed
+        # handling of scanning for unlocked files upon init . It was optimized etc
+        # and that message is no longer appears. yoh is not yet sure when it would
+        if external_versions['cmd:annex'] <= '8.20210428':
+            assert_in("for unlocked", cml.out)
 
 
 @with_tempfile
