+++ CONTRIBUTING.md
@@ -652,27 +652,33 @@ Refer datalad/config.py for information on how to add these environment variable
 - `update-changelog`: uses above `linkissues-changelog` and updates .rst changelog
 - `release-pypi`: ensures no `dist/` exists yet, creates a wheel and a source distribution and uploads to pypi.
 
-## Changelog section
+## Releasing with GitHub Actions, auto, and pull requests
 
-For the upcoming release use this template
+New releases of datalad are created via a GitHub Actions workflow built
+around [`auto`](https://github.com/intuit/auto).  Whenever a pull request is
+merged into `maint` that has the "`release`" label, `auto` updates the
+changelog based on the pull requests since the last release, commits the
+results, tags the new commit with the next version number, and creates a GitHub
+release for the tag.  This in turn triggers a job for building an sdist & wheel
+for the project and uploading them to PyPI.
 
-```markdown
-## 0.14.6 (??? ??, 2021) -- will be better than ever
+### Labelling pull requests
 
-bet we will fix some bugs and make a world even a better place.
+The section that `auto` adds to the changelog on a new release consists of the
+titles of all pull requests merged into master since the previous release,
+organized by label.  `auto` recognizes the following PR labels:
 
-### Major refactoring and deprecations
-
-- hopefully none
-
-### Fixes
-
-?
-
-### Enhancements and new features
-
-?
-```
+- `major` — for changes corresponding to an increase in the major version
+  component
+- `minor` — for changes corresponding to an increase in the minor version
+  component
+- `patch` — for changes corresponding to an increase in the patch/micro version
+  component; this is the default label for unlabelled PRs
+- `internal` — for changes only affecting the internal API
+- `documentation` — for changes only affecting the documentation
+- `tests` — for changes to tests
+- `dependencies` — for updates to dependency versions
+- `performance` — for performance improvements
 
 [link_zenodo]: https://github.com/datalad/datalad/blob/master/.zenodo.json
 [contrib_emoji]: https://allcontributors.org/docs/en/emoji-key
