### 2021-06-01 6df2c56
remove documentation of `major` label which we will not have for now
