### 2021-06-01 025b2bf
Fix add-label workflow syntax
