### 2021-06-02 37245a3
BF(TST): adjust test_version to match expectations from switching to auto
We will not bother any longer about __hardcoded_version__ and __full_version__:
versioneer produced version is a proper semver one, so let"s just use it alone
