### 2021-06-02 80d9a8b
RF/BF: determine version in _datalad_build_support via importlib_metadata
It will require datalad to be installed to do it correctly but IMHO it is
ok since e.g. for manpage generations we would need to have datalad
importable anyways
