+++ setup.py
@@ -38,6 +38,7 @@ requires = {
         'chardet>=3.0.4',      # rarely used but small/omnipresent
         'colorama; platform_system=="Windows"',
         'distro; python_version >= "3.8"',
+        'importlib-metadata; python_version < "3.8"',
         'iso8601',
         'humanize',
         'fasteners>=0.14',
