### 2021-06-02 6901cbb
BF: try to mitigate not listing of importlib-metadata in install requires for now
by delaying the import
