### 2021-06-02 9ce1367
Test for semver labels instead of trying to apply them
