+++ .github/workflows/add-label.yml
@@ -1,27 +1,0 @@
-name: Add "patch" or "minor" label to PR
-
-on:
-  pull_request:
-    type: [opened]
-
-jobs:
-  auto-release:
-    runs-on: ubuntu-latest
-    if: >
-      github.repository == 'datalad/datalad'
-      && !contains(github.event.pull_request.labels.*.name, 'patch')
-      && !contains(github.event.pull_request.labels.*.name, 'minor')
-      && !contains(github.event.pull_request.labels.*.name, 'major')
-    steps:
-      - name: Add label to PR
-        run: |
-          if [ "${{ github.event.pull_request.head.label }}" = "datalad:master" ]
-          then label=minor
-          else label=patch
-          fi
-          gh -R "$GITHUB_REPOSITORY" pr edit --add-label "$label" \
-            "${{ github.event.pull_request.number }}"
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-# vim:set sts=2:
