### 2021-06-03 748e546
New class CapturedException
Introducing this class to capture all textual info from an exception and
providing methods to provide formatted output from it, while holding no
references to the actual exception object - particularly its traceback
with frame references, that may create circular references.

It's intended to be used with for logging (and probably result reporting
later), which is why its output via __str__ can be enabled/disabled
completely by 'datalad.log.exc'. Thereby closes #4478.
