### 2021-06-03 41842a6
Improve/correct description of config var
