### 2021-06-03 041f14f
RF/TST: Turn exc_str into temporary shim
Intention is to replace current use of exc_str by using
CapturedException. However, not creating conflicts for now, and only
remove exc_str entirely after this was merged. Furthermore its usages in
result reporting is conceptually different and should get some extra
attention. Lastly, when replacing it, attention should be paid to the
purpose of `CapturedException` and not only change the call, but check
that we don't have the original exception object lingering around.

See #5716 for removing it entirely
