### 2021-06-03 7ed0ffe
TST: Add DATALAD_LOG_EXC=1 to CI runs
Since we test log output everywhere and this config now allows to
disable exception rendering in log output (which in accordance w/ #4478
is the new default), we need to enable it for the test runs.
