+++ .github/workflows/test_macos.yml
@@ -7,6 +7,9 @@ on:
 jobs:
   test:
     runs-on: macos-latest
+    env:
+      # The following is currently necessary as we test a lot of logging output in high detail:
+      DATALAD_LOG_EXC: 1
     strategy:
       fail-fast: false
       matrix:
