### 2021-06-03 0f586d3
RF+BF: centralize sanitization of reponame for github + sanitize full name
Previously sanitization (replacing " "s with "_" etc) was done only on
subdataset repository name.  It works fine until super-dataset repo
name with which subdataset is combined also has any of those "illegal"
characters which github would replace.

And our sanitization differs from how github does it -- github
replaces " "s with "-" whenever we replace with "_".

That lead to breaking the logic for "create-sibling-github" where we
first check if repository exists.  But if we do not sanitize
super-dataset repo name, it 
[…]
