### 2021-06-03 523beb4
BF: make hierarchies of create-sibling-github datasets install/get-able
This change introduces special handling for github URLs.  Since we do
have `create-sibling-github` I think it is Ok to also provide special
handling for such dataset hierarchies created on github (by DataLad)
within DataLad:

- before this change we were using POSIX-style composition for all URL
  + path, even though we know that it would not work for GitHub

- we do our "sanitization" of repository names we provide to GitHub
  upon create-sibling-github.  So we are the "central" place with
  that logic on sanitization (se
[…]
