### 2021-06-03 047c1cf
ENH(UX): state original purpose in NoDatasetFound exception + detail it for "get"
