### 2021-06-03 114b46e
ENH/RF: centralize eval_params defaults to their Parameter() definitions
Relation/difference in CLI invocation defaults is now explicitly
annotated.

Centralization:

- We can discover defaults via a simple loop, no need to move
  defaults specification to eval_defaults.

  This also helps to avoid coming up with per-param variables
  to contain the defaults

- argparser provides  default value via string interpolation, so
  we can just use %(default)s and %(default)r if so desired, and
  avoid referring to variables etc. Simplifies and harmonizes.

- a comment added about intended (IMHO) diff
[…]
