### 2021-06-07 0f6ec06
UX: When two or more clone URL templates are found, error out more gracefully
Previously, when someone would specify submodule clone candidates under the same name, we would crash
with an Attribute error. This is caught earlier now, fixing #5715, and the problematic candidate name
is reported to the user
