### 2021-06-07 ee76ff7
RF: Remove capture_exc in favor of constructor
classmethod turns out superfluous - there's only one way to get an
instance. Let that be the constructor.
