### 2021-06-07 985413a
BF: do not use .acquired - just get state from acquire()
Somewhat ad-hoc and makes an already confusing lock_if_check_fails even more so
by optionally changing its return signature. So eventually we might want
to RF this whole piece of code.  But I think it should work ok for now.
Anyways -- fasteners 0.16.x series might re-gain .acquired (see
https://github.com/harlowja/fasteners/issues/71) and we might want to see to RF
this PR properly if so instead of merging/releasing it right away.  But FWIW
Closes #5717
