### 2021-06-07 b5015cc
Add test for number of resulting commits
Changes to .gitmodules on clone should be included in
the commit that adds the subdatasets.
