### 2021-06-07 3efe198
foreach: --state='installed' to be the default
