+++ datalad/local/subdatasets.py
@@ -14,6 +14,7 @@ __docformat__ = 'restructuredtext'
 import logging
 import re
 import os
+import warnings
 
 from datalad.interface.base import Interface
 from datalad.interface.utils import eval_results
@@ -22,11 +23,13 @@ from datalad.interface.results import get_status_dict
 from datalad.support.constraints import (
     EnsureStr,
     EnsureNone,
+    NoneDeprecated,
 )
 from datalad.support.param import Parameter
 from datalad.support.exceptions import CommandError
 from datalad.interface.common_opts import (
     contains,
+    dataset_state,
     fulfilled,
     recursion_flag,
     recursion_limit,
@@ -157,6 +160,7 @@ class Subdatasets(Interface):
             a dataset method PY].""",
             nargs='*',
             constraints=EnsureStr() | EnsureNone()),
+        state=dataset_state,
         fulfilled=fulfilled,
         recursive=recursion_flag,
         recursion_limit=recursion_limit,
@@ -199,13 +203,35 @@ class Subdatasets(Interface):
     def __call__(
             path=None,
             dataset=None,
-            fulfilled=None,
+            state='any',
+            fulfilled=NoneDeprecated,
             recursive=False,
             recursion_limit=None,
             contains=None,
             bottomup=False,
             set_property=None,
             delete_property=None):
+        if fulfilled is not NoneDeprecated:
+            # the two mirror options do not agree and the deprecated one is
+            # not at default value
+            warnings.warn("subdatasets's `fulfilled` option is deprecated "
+                          "and will be removed in a future release, "
+                          "use the `state` option instead.",
+                          DeprecationWarning)
+            if state != 'any':
+                raise ValueError("Do not specify both 'fulfilled' and 'state', use 'state'")
+            # honor the old option for now
+            state = {
+                None: 'any',
+                True: 'present',
+                False: 'absent',
+            }[fulfilled]
+        # Path of minimal resistance/code-change - internally here we will reuse fulfilled
+        fulfilled = {
+            'any': None,
+            'present': True,
+            'absent': False,
+        }[state]
         ds = require_dataset(
             dataset, check_installed=True, purpose='subdataset reporting/modification')
 
