+++ datalad/distribution/tests/test_install.py
@@ -320,7 +320,7 @@ def test_install_recursive(src, path_nr, path_r):
         ok_(not sub.is_installed(),
             "Unintentionally installed: %s" % (sub,))
     # this also means, subdatasets to be listed as not fulfilled:
-    eq_(set(ds.subdatasets(recursive=True, fulfilled=False, result_xfm='relpaths')),
+    eq_(set(ds.subdatasets(recursive=True, state='absent', result_xfm='relpaths')),
         {'subm 1', '2'})
 
     # now recursively:
@@ -351,7 +351,7 @@ def test_install_recursive(src, path_nr, path_r):
                                                  eval_availability=True)
         assert_false(any(st["has_content"] for st in ainfo.values()))
     # no unfulfilled subdatasets:
-    ok_(top_ds.subdatasets(recursive=True, fulfilled=False) == [])
+    ok_(top_ds.subdatasets(recursive=True, state='absent') == [])
 
     # check if we can install recursively into a dataset
     # https://github.com/datalad/datalad/issues/2982
@@ -479,8 +479,8 @@ def test_install_known_subdataset(src, path):
     # subdataset not installed:
     subds = Dataset(opj(path, 'subm 1'))
     assert_false(subds.is_installed())
-    assert_in('subm 1', ds.subdatasets(fulfilled=False, result_xfm='relpaths'))
-    assert_not_in('subm 1', ds.subdatasets(fulfilled=True, result_xfm='relpaths'))
+    assert_in('subm 1', ds.subdatasets(state='absent', result_xfm='relpaths'))
+    assert_not_in('subm 1', ds.subdatasets(state='present', result_xfm='relpaths'))
     # install it:
     ds.install('subm 1')
     ok_(subds.is_installed())
@@ -489,8 +489,8 @@ def test_install_known_subdataset(src, path):
     # new repository initiated
     eq_(set(subds.repo.get_indexed_files()),
         {'test.dat', 'INFO.txt', 'test-annex.dat'})
-    assert_not_in('subm 1', ds.subdatasets(fulfilled=False, result_xfm='relpaths'))
-    assert_in('subm 1', ds.subdatasets(fulfilled=True, result_xfm='relpaths'))
+    assert_not_in('subm 1', ds.subdatasets(state='absent', result_xfm='relpaths'))
+    assert_in('subm 1', ds.subdatasets(state='present', result_xfm='relpaths'))
 
     # now, get the data by reinstalling with -g:
     ok_(subds.repo.file_has_content('test-annex.dat') is False)
@@ -777,7 +777,7 @@ def test_install_consistent_state(src, dest, dest2, dest3):
 
     def check_consistent_installation(ds):
         datasets = [ds] + list(
-            map(Dataset, ds.subdatasets(recursive=True, fulfilled=True,
+            map(Dataset, ds.subdatasets(recursive=True, state='present',
                                         result_xfm='paths')))
         assert len(datasets) == 2  # in this test
         for ds in datasets:

[…]
