### 2021-06-08 e4253d1
WiP Q&D AnnexRepo.add_ batched (by default) if more than a single file passed
