### 2021-06-08 48d0562
WiP Q&D AnnexRepo.add_ batched (by default) if more than a single file passed
