+++ datalad/support/tests/test_annexrepo.py
@@ -2384,7 +2384,7 @@ def check_files_split(cls, topdir):
 @known_failure_windows  # does not find files to add (too long paths?)
 @slow  # 313s  well -- if errors out - only 3 sec
 def test_files_split():
-    for cls in GitRepo, AnnexRepo:
+    for cls in AnnexRepo, :
         yield check_files_split, cls
 
 
