### 2021-06-08 8002107
WiP Q&D AnnexRepo.add_ batched (by default) if more than a single file passed
