### 2021-06-08 4479b1b
WiP Q&D AnnexRepo.add_ batched (by default) if more than a single file passed
