### 2021-06-08 2f974d5
RF: Centralize with the duplicated invocation within _save_add
