### 2021-06-08 0ddc05c
RF: Centralize with the duplicated invocation within _save_add
