### 2021-06-08 5f4fb8f
RF: Centralize with the duplicated invocation within _save_add
