### 2021-06-08 fc6c2b2
RF: Centralize with the duplicated invocation within _save_add
