### 2021-06-09 7c8cb62
BF/TST: Query the DEFAULT_BRANCH instead of adjusted one
