### 2021-06-09 datalad/datalad#5685 @adswa
hey @jkanche, just a quick ping on this. I feel like the presence of a license in a text file with a commonly used name should suffice, but we could also ask the editor in the JOSS issue about this.
