### 2021-06-09 9ac77fa
ENH: add newly created README files into Git.
This configures the dataset to place a newly create README file into Git
and adds the new parameter --annex to add-readme (False by default),
which allows to override this default.
Fixes #5713.

Better fit the parameter name
