### 2021-06-09 a8689ac
TST/MNT: Adjust test for updated README template
