### 2021-06-09 1a3ad7d
UX: When a remote is specified as a path, provide a hint about --to
Fixes 5712
