### 2021-06-10 aa800c7
Add config datalad.remote.archives.autoclean
This flag determines whether or not the datalad-archives
special remote keeps extracted files under .git as an
extraction cache. Default set to False, since this can
be expensive to keep around, it's not obvious for users
that it's there and how to remove it, and generally the
defaults should scale. True doesn't scale here.
