### 2021-06-10 73cbe11
WIP: alternative approach to report Key errors in subdataset clone url templates
