### 2021-06-10 3b927e3
ENH: harmonize "purpose" statements to imperative form
We had a mix, and even though "ing form was probably more prevalent, I felt that shorter
imperative form would be better, even though it would be not "for {purpose}" but
"to {purpose}" if to be displayed. (as eg in #5708)
