### 2021-06-10 13ea4fb
ENH(TST): populate heavy tree with 100 unique keys (not just 1) among 10,000
