### 2021-06-10 e24063f
RF: Reduce result keeping to what's needed
