### 2021-06-10 f11a3fe
ENH(TST,WIN): reenable test_files_split on leaner tree on windows
It still managed to hang my VM running out of the space.  So I will
see if appveyor CI handles it better.
