### 2021-06-10 2871061
ENH: Do not condition the decision on annexing a readme on the existence of the .gitattributes file
