### 2021-06-10 5654727
BF: default _n_auto_jobs before we might call _init
Otherwise _call_annex would fail since _n_auto_jobs would not yet be
assigned.  We never hit this bug because we default to jobs=None and
it was only for higher level interfaces to pass anything by None, and
by then AnnexRepo would be fully instantiated and thus _n_auto_jobs
assigned.
