### 2021-06-10 e3e75aa
ENH: set jobs='auto' (instead of None) for annex interfaces which support --jobs
Thus in effect we would actually make  datalad.runtime.max-annex-jobs
used, and its doc reflective of what is actually happening:

   'title': 'Maximum number of git-annex jobs to request when "jobs" option set to "auto" (default)'
