### 2021-06-10 bf066c7
RF/BF: do not mask annex.jobs of annex by default, and allow for 0 (do not set --jobs)
otherwise
- we cannot overload the behavior to have only 1 job (`if` was
adjusted)
- we cannot overload the behavior to say "do not overload git-annex
configuration"
