+++ .travis.yml
@@ -17,6 +17,8 @@ env:
     - DATALAD_TESTS_SSH=1
     - DATALAD_LOG_ENV=GIT_SSH_COMMAND
     - TESTS_TO_PERFORM=datalad
+    # To potentially speed up testing, let's do annex jobs in parallel by default
+    - DATALAD_RUNTIME_MAX__ANNEX__JOBS=5
     # Should be an array, travis breaks on it, define/adjust in the "before_install"
     #- NOSE_OPTS=( -v )
     # Note, that there's "turtle" as well, which is always excluded from
@@ -38,6 +40,8 @@ matrix:
     # We cannot have empty -A selector, so the one which always will be fulfilled
     - NOSE_SELECTION=
     - NOSE_SELECTION_OP=not
+    # A single run where we do not instruct annex on number of jobs
+    - DATALAD_RUNTIME_MAX__ANNEX__JOBS=0
     # To test https://github.com/datalad/datalad/pull/4342 fix.
     # From our testing in that PR seems to have no effect, but kept around since should not hurt.
     - LC_ALL=ru_RU.UTF-8
@@ -46,6 +50,8 @@ matrix:
     env:
     - NOSE_SELECTION=
     - NOSE_SELECTION_OP=not
+    # A single run where we instruct annex to just have a single job
+    - DATALAD_RUNTIME_MAX__ANNEX__JOBS=1
   - if: type = cron
     python: 3.6
     # Single run for Python 3.6
