### 2021-06-10 8e5398a
RF+ENH: move "auto" logic into setting the default for max-annex-jobs
- this centralizes the setting, and makes it so that jobs='auto' is
  indeed gaining some parallelization.  Otherwise we would default to
  0 (what is explicitly configured for annex), or 1 (no parallelization).

- simplifies all the "mind logic" on the effect of having 'auto' set.
