### 2021-06-10 915ec6c
tributors: annotate commit and PR with [skip ci]
It seems that now even github should support it if marker is in the commit
message: https://github.blog/changelog/2021-02-08-github-actions-skip-pull-request-and-push-workflows-with-skip-ci/
