### 2021-06-11 #3262 @mih
Hey! Thanks in advance to the reviewers for wading through our code and docs -- certainly not a quick thing. I would nevertheless like to ask if there is anything we can do to streamline the process. I am only aware of one issue having been reported https://github.com/datalad/datalad/issues/5685

Have @jkanche @szorowi1 found something that we could proactively tackle?

Thx and best!
