### 2021-06-11 349ee21
ENH: Add --dry-run to clean
With this switch `clean` reports on existing temporary locations, that
it is able to clean up. This modes yields result records with
'action="clean [dry-run]"' and thereby is inline with `run --dry-run`.
