### 2021-06-14 c60eac4
Confine number of paths to one when checking if a path is a sibling
Co-authored-by: Michael Hanke <michael.hanke@gmail.com>
