### 2021-06-14 cfe311c
improve hint message
