+++ datalad/core/distributed/push.py
@@ -273,8 +273,8 @@ class Push(Interface):
                     p for p in ensure_list(path) if p in sr
                 ]
             if potential_remote:
-                msg = "It seems like you specified a sibling name {} " \
-                      "as a path. Forgot to use --to?".format(potential_remote)
+                msg = "{} matches a sibling name and not a path. " \
+                      "Forgot --to?".format(potential_remote)
                 yield dict(
                     res_kwargs,
                     status='impossible',
