### 2021-06-14 5ede193
Reuse the 'hints' setup in result rendering of push's custom result renderer
