### 2021-06-14 7d12d76
UX: Render hints after result reporting, and shorten hint message
adjust test
