### 2021-06-14 a46f21e
When no path matches, but it resembles a sibling, return early
