### 2021-06-14 e9f2fc9
Make lock file location configurable
This introduces `datalad.locations.locks` to specify the location to be
used for (global) lock files. Defaults to the currently used
`<user cache dir>/locks`.

Closes #5706
