### 2021-06-14 f53c74f
DOC: advise on clean .gitattributes file state
