### 2021-06-15 ce85e09
Make cmd mandatory positional argument
I left a check for it to not be empty or None since it
should not urt and could prevent some incorrect
invocations with a meaningful message
