### 2021-06-16 72eb760
ENH: foreach - support native python functions
Would make foreach useful for direct use within Python code, e.g.
as a convenience layer instead of using ProducerConsumer pattern
