### 2021-06-16 datalad/datalad#5685 @arokem
From the point of view of the JOSS review, it's fine for it to be where it is.

@jkanche : can this issue be closed? Does this check your "License" box in the review?
