### 2021-06-16 da80745
NF: Clone URL mapping feature
It allows for configuring an arbitrary number of URL mapping
specification. With this change one can configure datalad to transfer
URLs like https://osf.io/q8xnk/ (copied from a browser window), which do
not work, but can be reliably mapped into working URLs (osf://q8xnk in
this case) via a regular expression specification.

Fixes gh-4822
