### 2021-06-18 e66e185
RF(TST): disable test on not consuming all produced items
