### 2021-06-18 9a72b34
BF: do not "echo -e" into GITHUB_ENV, dedent [skip ci] in commit msg
with echo -e we have workflow failing with

    Error: Unable to process file command 'env' successfully.
    Error: Invalid environment variable format '[skip ci]''

so let's see if this helps

[ci skip]
