### 2021-06-18 #3262 @szorowi1
Very sorry for the delay! I've worked through my review.

It's a very impressive piece of software. I've read through much of the documentation and several of the helpful tutorial videos. I haven't been able to test every aspect of its functionality, but I've successfully been able to install Datalad and work through the tutorials. I especially like the conversational style of the docs -- makes it very easy to read and follow.

I have no quibbles. It's very clear how much effort has gone into the documentation and software to make it user-friendly and easy-to-use.
