### 2021-06-23 9138af0
BF: Prefer committed subdataset location info over guesses
gh-5033 describes a situation where the installation of a sub-subdataset
fails due to the particular condition of the _checkout_ the subdataset
was cloned from. The issues originally expressed the desire for a more
in-depth exploration of possible superdatasets and their subdataset
configurations. However, the undesirable failure can also be avoided by
simply preferring the location information that is committed in each
superdataset, at all clone iterations on the way to a target subdataset.

Effectively, the change reorders the locati
[…]
