+++ datalad/distribution/tests/test_get.py
@@ -104,20 +104,20 @@ def test_get_flexible_source_candidates_for_submodule(t, t2, t3):
     # own location default remote for current branch
     clone_subpath = str(clone.pathobj / 'sub')
     eq_(f(clone, dict(path=clone_subpath, parentds=clone.path)),
-        [dict(cost=500, name=DEFAULT_REMOTE, url=ds_subpath)])
+        [dict(cost=650, name=DEFAULT_REMOTE, url=ds_subpath)])
     eq_(f(clone, dict(path=clone_subpath, parentds=clone.path, gitmodule_url=sshurl)),
-        [dict(cost=500, name=DEFAULT_REMOTE, url=ds_subpath),
-         dict(cost=600, name=DEFAULT_REMOTE, url=sshurl)])
+        [dict(cost=600, name=DEFAULT_REMOTE, url=sshurl),
+         dict(cost=650, name=DEFAULT_REMOTE, url=ds_subpath)])
     eq_(f(clone, dict(path=clone_subpath, parentds=clone.path, gitmodule_url=httpurl)),
-        [dict(cost=500, name=DEFAULT_REMOTE, url=ds_subpath),
-         dict(cost=600, name=DEFAULT_REMOTE, url=httpurl)])
+        [dict(cost=600, name=DEFAULT_REMOTE, url=httpurl),
+         dict(cost=650, name=DEFAULT_REMOTE, url=ds_subpath)])
 
     # make sure it does meaningful things in an actual clone with an actual
     # record of a subdataset
     clone_subpath = str(clone.pathobj / 'sub')
     eq_(f(clone, clone.subdatasets(return_type='item-or-list')),
         [
-            dict(cost=500, name=DEFAULT_REMOTE, url=ds_subpath),
+            dict(cost=600, name=DEFAULT_REMOTE, url=ds_subpath),
     ])
 
     # check that a configured remote WITHOUT the desired submodule commit
@@ -126,7 +126,7 @@ def test_get_flexible_source_candidates_for_submodule(t, t2, t3):
                    result_renderer='disabled')
     eq_(f(clone, clone.subdatasets(return_type='item-or-list')),
         [
-            dict(cost=500, name=DEFAULT_REMOTE, url=ds_subpath),
+            dict(cost=600, name=DEFAULT_REMOTE, url=ds_subpath),
     ])
     # inject a source URL config, should alter the result accordingly
     with patch.dict(
@@ -134,7 +134,7 @@ def test_get_flexible_source_candidates_for_submodule(t, t2, t3):
             {'DATALAD_GET_SUBDATASET__SOURCE__CANDIDATE__BANG': 'youredead'}):
         eq_(f(clone, clone.subdatasets(return_type='item-or-list')),
             [
-                dict(cost=500, name=DEFAULT_REMOTE, url=ds_subpath),
+                dict(cost=600, name=DEFAULT_REMOTE, url=ds_subpath),
                 dict(cost=700, name='bang', url='youredead', from_config=True),
         ])
     # we can alter the cost by given the name a two-digit prefix
@@ -144,7 +144,7 @@ def test_get_flexible_source_candidates_for_submodule(t, t2, t3):
         eq_(f(clone, clone.subdatasets(return_type='item-or-list')),
             [
                 dict(cost=400, name='bang', url='youredead', from_config=True),
-                dict(cost=500, name=DEFAULT_REMOTE, url=ds_subpath),
+                dict(cost=600, name=DEFAULT_REMOTE, url=ds_subpath),
         ])
     # verify template instantiation works
     with patch.dict(
@@ -152,7 +152,7 @@ def test_get_flexible_source_candidates_for_submodule(t, t2, t3):
[…]
