### 2021-06-24 #3262 @yarikoptic
Thank you @szorowi1 @jkanche and @arokem !  Meanwhile, we ran into a paper which might be giving even a better presentation to the DataLad that we had prepared for JOSS https://psyarxiv.com/w8trm .  I assume we still have a possibility to adjust the manuscript with more references?
