+++ paper.bib
@@ -173,3 +173,13 @@
 	title = {The brain imaging data structure, a format for organizing and describing outputs of neuroimaging experiments},
 	journal = {Scientific Data}
 }
+
+@article{Gautheron_2021,
+	doi = {10.31234/osf.io/w8trm},
+	url = {https://doi.org/10.31234/osf.io/w8trm},
+	year = 2021,
+	month = {may},
+	publisher = {Center for Open Science},
+	author = {Lucas Gautheron and Nicolas Rochat and Alejandrina Cristia},
+	title = {Managing, storing, and sharing long-form recordings and their annotations}
+}
