+++ paper.bib
@@ -183,3 +183,19 @@
 	author = {Lucas Gautheron and Nicolas Rochat and Alejandrina Cristia},
 	title = {Managing, storing, and sharing long-form recordings and their annotations}
 }
+
+@article{jtrack,
+  author    = {Mehran Sahandi Far and
+               Michael Stolz and
+               Jona M. Fischer and
+               Simon B. Eickhoff and
+               Juergen Dukart},
+  title     = {JTrack: {A} Digital Biomarker Platform for Remote Monitoring in Neuropsychiatric
+               and Psychiatric Diseases},
+  journal   = {CoRR},
+  volume    = {abs/2101.10091},
+  year      = {2021},
+  url       = {https://arxiv.org/abs/2101.10091},
+  archivePrefix = {arXiv},
+  eprint    = {2101.10091},
+}
