+++ paper.bib
@@ -11,6 +11,19 @@
 	journal = {Neuroforum}
 }
 
+@software{datalad-ukb:zenodo,
+  author       = {Hanke, Michael and
+                  Waite, Laura K. and
+                  Poline, Jean-Baptiste and
+                  Hutton, Alexandre},
+  title        = {DataLad extension for working with the UKbiobank},
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {0.3.3},
+  doi          = {10.5281/zenodo.4773629},
+  url          = {https://doi.org/10.5281/zenodo.4773629}
+}
+
 @misc{datalad-container:zenodo,
   doi = {10.5281/ZENODO.2431914},
   url = {https://zenodo.org/record/2431914},
