### 2021-06-25 #3262 @mih
I have now updated the manuscript with the following changes:

- spelling fixes
- add citation for another system that uses datalad
- a highres version of figure 2
- list the datalad UKBiobank extension too

It should hopefully be acceptable for publication now. Thx for given us the opportunity for these last-minute updates.
