### 2021-06-25 99ea5c1
RF: Adjust test for subdataset clone from .gitmodule URL by default
Besides explicitly setting up the superdataset with relative submodule
URLs this also required tuning some test conditions, because the
submodule addition is not longer the last change on the branch.
