### 2021-06-26 785ac0e
Change classification of BIDS paper
The current bibliography item that gets rendered into the article is a bit wonky (including the publisher name, for example). I think that reclassifying this explicitly as an article may make the rendered bibliography item more conformant.
