+++ paper.bib
@@ -174,7 +174,7 @@
     year={2021}
 }
 
-@misc{bids:paper,
+@article{bids:paper,
 	doi = {10.1038/sdata.2016.44},
 	url = {https://doi.org/10.1038/sdata.2016.44},
 	year = 2016,
