### 2021-06-26 #3262 @arokem
Hello! Overall, looks good. I have a couple of small suggestions:

1. I think that the section title "Overview of the DataLad and its ecosystem"  should be "Overview of DataLad and its ecosystem" or "Overview of the DataLad software and its ecosystem" (I would prefer the former), unless there is some intention in this wording that I am missing.

2. https://github.com/datalad/datalad/pull/5766
