### 2021-06-26 #3262 @yarikoptic
Thank you @arokem!
1. took the former (removing "the")
2. included the change to bib - not sure how it came to be `@misc` ;)
