### 2021-06-26 #3262 @arokem
OK! At this point could you:

- [ ] Make a tagged release of your software and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

We can then move forward with a final check by one of the EiC, before accepting the paper.
