### 2021-06-26 #3262 @yarikoptic
Just to make sure - paper branch can remain detached (not merged into) from a release branch (maint) and tag, correct?
