### 2021-06-26 #3262 @arokem
I believe that's correct, but let's bring in @openjournals/joss-eics to help resolve this.

This is a case where the paper is on a separate branch. The release version that gets tagged and archived does not have to include this branch in it, does it?
