### 2021-06-26 #3262 @arfon
> Just to make sure - paper branch can remain detached (not merged into) from a release branch (maint) and tag, correct?

Yes, that's right. We can accept the submission without the paper actually _landing_ in the merged code into the repository.
