### 2021-06-26 #3262 @yarikoptic
> * Check that the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it).

Does it require to have author list in the same **order** in this paper and on Zenodo, or just that the "set" of authors match?
