+++ docs/source/changelog.rst
@@ -2,6 +2,25 @@
 
 Change log
 **********
+0.14.6 (Sun Jun 27 2021)
+========================
+
+Internal
+--------
+
+-  BF: update changelog conversion from .md to .rst (for sphinx)
+   `#5757 <https://github.com/datalad/datalad/pull/5757>`__
+   (`@yarikoptic <https://github.com/yarikoptic>`__
+   `@jwodder <https://github.com/jwodder>`__)
+
+Authors: 2
+----------
+
+-  John T. Wodder II (`@jwodder <https://github.com/jwodder>`__)
+-  Yaroslav Halchenko (`@yarikoptic <https://github.com/yarikoptic>`__)
+
+--------------
+
 0.14.5 (Mon Jun 21 2021)
 ========================
 
@@ -88,6 +107,8 @@ Pushed to ``maint``
 -------------------
 
 -  MNT: Post-release dance (`@kyleam <https://github.com/kyleam>`__)
+
+.. _internal-1:
 
 Internal
 --------
@@ -425,7 +446,7 @@ Enhancements and new features
    default, has been improved. A core piece of the new approach is
    registering the commit of the primary branch, not its checked out
    adjusted branch, in the superdataset. Note: This means that
-   ``git   status`` will always consider a subdataset on an adjusted
+   ``git status`` will always consider a subdataset on an adjusted
    branch as dirty while ``datalad status`` will look more closely and
    see if the tip of the primary branch matches the registered commit.
    (`#5241 <https://github.com/datalad/datalad/issues/5241>`__)
@@ -554,10 +575,10 @@ Enhancements and new features
 
 -  The ``ConfigManager`` methods ``get``, ``getbool``, ``getfloat``, and
    ``getint`` now return a single value (with same precedence as
-   ``git   config --get``) when there are multiple values for the same
-   key (in the non-committed git configuration, if the key is present
-   there, or in the dataset configuration). For ``get``, the old
-   behavior can be restored by specifying ``get_all=True``.
+   ``git config --get``) when there are multiple values for the same key
+   (in the non-committed git configuration, if the key is present there,
+   or in the dataset configuration). For ``get``, the old behavior can
+   be restored by specifying ``get_all=True``.
    (`#4924 <https://github.com/datalad/datalad/issues/4924>`__)
 
 -  Command-line scripts are now defined via the ``entry_points``
@@ -1025,7 +1046,7 @@ Enhancements and new features
 
 -  ``datalad push`` now avoids unnecessary ``git push`` dry runs and
    pushes all refspecs with a single ``git push`` call rather than
-   invoking ``git   push`` for each one.
+   invoking ``git push`` for each one.
    (`#4692 <https://github.com/datalad/datalad/issues/4692>`__)
    (`#4675 <https://github.com/datalad/datalad/issues/4675>`__)
 
@@ -1378,9 +1399,8 @@ Fixes
 -  A longstanding regression in argcomplete-based command-line
    completion for Bash has been fixed. You can enable completion by
    configuring a Bash startup file to run
-   ``eval   "$(register-python-argcomplete datalad)"`` or source
-   DataLad’s ``tools/cmdline-completion``. The latter should work for
-   Zsh as well.
+   ``eval "$(register-python-argcomplete datalad)"`` or source DataLad’s
+   ``tools/cmdline-completion``. The latter should work for Zsh as well.
    (`#4477 <https://github.com/datalad/datalad/issues/4477>`__)
 
 -  `publish <http://datalad.readthedocs.io/en/latest/generated/man/datalad-publish.html>`__
@@ -1461,13 +1481,12 @@ Fixes
 
 -  The default for the ``--jobs`` option, “auto”, instructed DataLad to
[…]
