### 2021-06-28 d2b6914
NF: Institutionalize GitHub subdataset URL mapping
This is an alternative fix to datalad/datalad#5709 that makes dataset
hierarchies installable from GitHub -- as they are put up by
`create-sibling-github`. The main difference here is that there is
no service-specific code needed, only a configuration of the necessary
mapping(s). Consequently, the mechanism is applicable to other types of
mappings and other services.

The GitHub usecase is already rather sophisticated, but I believe the
approach of an initial critical match (to enter a mapping series), plus
an arbitrary number of additional ma
[…]
