### 2021-06-28 #3262 @yarikoptic
While revisiting the list of the requirements for Zenodor archive entry, another concern is

> ... make sure the list is correct and people who only made a small fix are not on it

- we have contributors who contributed no "fix" at all, but were significant and notable contributors to see to the success of the project.  I think this requirement/statement should be adjusted for JOSS to not suggest a demand only for major "in commit/code" contributions: projects do often need administrators, mentors, user support,   and other "personnel" which would not be submitting a "fix" but would be critical.
- our zenodo entry might have (didn't check) minor "small fix" contributors to the software which decided to not become a co-author to the paper (and rather just to be acknowledged in Acknowledgements section).  I would hate to remove them from the list of authors on Zenodo.

So I hope that it would be acceptable to have Zenodo list of authors to be a superset of the authors for the publication and not necessarily matching the order.

If not acceptable: I guess we can mint a single release just for the purpose of JOSS publication with the desired listing/order.

Thank you in advance for the clarification(s). (attn @arokem @arfon )
