### 2021-06-29 #3262 @arfon
> So I hope that it would be acceptable to have Zenodo list of authors to be a superset of the authors for the publication and not necessarily matching the order.

Yes this is OK. We absolutely _must_ have an archive of the software to associate with the review, and we _prefer_ that the archive metadata (authors, title etc.) match the JOSS paper. This later requirement though is a soft one so it's OK to proceed.
