### 2021-06-29 19278eb
NF: Introduce mapping for OSF URLs
As noted in the code, this should move to the datalad-osf extension,
as soon as it could provide such configuration.
https://github.com/datalad/datalad/issues/5769

But for now, it can meaningfully serve as a demo how multiple
subsitution series can coexist.
