### 2021-06-29 3864a53
Tune documentation to describe implementation more precisely
Leave a comment for a potential future sorting feature.
