+++ setup.py
@@ -48,6 +48,7 @@ requires = {
     ],
     'publish': [
         'PyGithub',          # nice to have
+        'python-gitlab',     # required for create-sibling-gitlab
     ],
     'misc': [
         'pyperclip',         # clipboard manipulations
