### 2021-06-29 #3262 @yarikoptic
We have been meticulously archiving datalad on zenodo (and pypi and debian and neurodebian and their snapshots repos) through the years and https://zenodo.org/record/5034875#.YNsmFXWYXjE is the one for recent `datalad/datalad: 0.14.6`.  Would it be sufficient (given the relaxed *prefer* for the metadata)?
