### 2021-07-01 #3262 @yarikoptic
Thank you @arfon @szorowi1, @jkanche and the last but not least @arokem for making this mundane Thursday a "once in a life time" special occasion ;)
The idea for the "DataLad paper" was painful (we exercised a number of concepts to pursue), and the goals, format and structure of JOSS publication and review process helped us tremendously to finally make it happen!

FWIW, if might come handy for some, https://github.com/datalad/datalad-git-bug-dumps/tree/master/tools provides some extra helper we used (on top of git-bug dump) to decide on co-authors etc to invite!  No JOSS paper is expected to come for that ;)  But may be it could give ideas to establish some JOSS-tools collection of helpers on how to objectively figure out the list of co-authors to invite etc.
